Changes to Linux platform settings
I've decided to make some adjustments to the platform settings after
trying to build RPM, DEB, and Flatpak packages. The various packaging
tools rely on being able to set their own optimization, debugging, and
hardening flags via the CFLAGS and LDFLAGS environment variables, so this
brings in that capability.

Additionally, building in a Flatpak environment requires explicitly
linking with the -pthread flag. I'm not sure why this was implicit before,
but I've added it to the linker macro, as it doesn't hurt to specify it
explicitly in any case.